feat(nvim): manage all LSPs via nvim-lsp-setup
I did like using the installer as a single point of management, but the
author is planning to deprecate automation and rely more on those
wrappers. So the only responsibility of the installer would be to
install LSPs, but not set them up.

This change is very fresh and nvim-lsp-setup has not transitioned to the
new way of installing LSPs but the abstraction should allow doing so
when the old way is officially deprecated.

Also added some jsonls schemas