feat(nvim): automatically fetch json schemas
Instead of hardcoded json schemas this solution automatically fetches
the JSON Schema Catalog from the SchemaStore API and caches them. They
are then loaded when the language server is set up.

There is also a command introduced `:RefreshJsonSchemas` to update the
JSON Schema Catalog file. This probably doesn't need to update that
often.