WIP - Waiting for NeoVim 0.8
Today every time I save a file that has more than one language server
(which in NullLS case would be most of them), it asks which one to use
for formatting and that's annoying as hell. This PR uses the NeoVim 0.8
API but I'm not ready to switch to using HEAD for day to day work.